split up interop project from core project
