+++ modules/rocksdb-interop/src/nativeInterop/cinterop/rocksdb.def
