+++ modules/rocksdb-interop/src/nativeMain/kotlin/generatedOptions.kt
