+++ settings.gradle.kts
@@ -2,6 +2,7 @@ rootProject.name = "kotlin-on-the-rocksdb"
 
 include(
   ":modules:rocksdb",
+  ":modules:rocksdb-interop",
 )
 
 enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
